final commit.

+++ SimpleGit^1.0/src/core/JitReset.java
@@ -1,7 +1,6 @@
 /*该类实现回滚功能，主要分为3个模式：soft，mixed和hard。
 commitId指的是key值。
 */
-
 package core;
 import fileoperation.FileCreation;
 import fileoperation.FileDeletion;
@@ -9,10 +8,13 @@ import fileoperation.FileReader;
 import gitobject.*;
 import repository.Repository;
 import java.io.*;
+import java.util.HashMap;
+import java.util.HashSet;
 
 
 public class JitReset{
     public static void reset(String mode,String commitID) {
+
         if(FileReader.objectExists(commitID)){
             //根据commitID生成一个commit类
             Commit com = Commit.deserialize(commitID);
@@ -23,7 +25,7 @@ public class JitReset{
             ORIG_HEAD orig_head = new ORIG_HEAD(Head.getCurrentCommit());
             orig_head.compressWrite();
 
-            if(mode =="soft"){
+            if(mode.equals("soft")){
                 try{
                     //更新head指针所指向的分支
                     Branch curbranch = Branch.getCurBranch();
@@ -33,7 +35,7 @@ public class JitReset{
                     e.printStackTrace();
                 }
             }
-            if(mode=="mixed"){
+            if(mode.equals("mixed")){
                 try{
                     //更新head指针所指向的分支
                     Branch curbranch = Branch.getCurBranch();
@@ -46,7 +48,7 @@ public class JitReset{
                     e.printStackTrace();
                 }
             }
-            if(mode=="hard"){
+            if(mode.equals("hard")){
                 try{
                     //更新head指针所指向的分支
                     Branch curbranch = Branch.getCurBranch();
@@ -58,21 +60,41 @@ public class JitReset{
 
                     //清空工作区所有文件，把commitID中的内容还原出来。
                     File[] worktree = new File(Repository.getWorkTree()).listFiles();
-                    for(File f:worktree){
+                    // 已被跟踪的文件
+                    Status status = Status.deserialize();
+                    HashSet<String> trackedSet = status.getTrackedSet();
+
+                    for(File f : worktree){
+                        if (!trackedSet.contains(f.getName()) || f.getName().equals(".jit")) continue;
                         FileDeletion.deleteFile(f);
                     }
+
                     FileCreation.recoverWorkTree(indextree, Repository.getWorkTree());
                 }catch(Exception e){
                     e.printStackTrace();
                 }
     
             }
+
+            System.out.println("HEAD is now at " + commitID.substring(0,7) + " " + com.getMessage());
+
         }else{
+
             System.out.println(commitID+"does not exist.");
+
         }
         
 
     }
+
+    public static void resetOneStep(String mode) {
+        Commit com = Commit.deserialize(Head.getCurrentCommit());
+        String parent_key = com.getParent();
+        if (mode.equals("soft") || mode.equals("mixed") || mode.equals("hard")) {
+            reset(mode, parent_key);
+        }
+    }
+
 
 
 }
